# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Credence API
  description: Generated OpenAPI documentation from Zod schemas
servers:
  - url: https://api.credence.org/v1
components:
  schemas:
...
    SlashRequest:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4e5f6a7b8
        targetAddress:
          type: string
          example: GDUKMGUGDZQK6YHYA5Z6AY2G4XDSZPSZ3SW5UN3ARVMO6QSRDWP5YLEX
        reason:
          type: string
          example: Repeated SLA violations on delivery commitments
        requestedBy:
          type: string
          example: validator-12
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T09:00:00.000Z
        votes:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
        status:
          $ref: "#/components/schemas/SlashRequestStatus"
        threshold:
          type: integer
          example: 3
        totalSigners:
          type: integer
          example: 5
      required:
        - id
        - targetAddress
        - reason
        - requestedBy
        - createdAt
        - votes
        - status
        - threshold
        - totalSigners
    Vote:
      type: object
      properties:
        voterId:
          type: string
          example: validator-3
        choice:
          $ref: "#/components/schemas/VoteChoice"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp the vote was cast
          example: 2024-01-15T10:00:00.000Z
      required:
        - voterId
        - choice
        - timestamp
    VoteChoice:
      type: string
      enum:
        - approve
        - reject
    SlashRequestStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
    GovernanceError:
      type: object
      properties:
        error:
          type: string
          example: NotFound
        message:
          type: string
          example: Slash request not found
      required:
        - error
        - message
    CreateSlashRequestBody:
      type: object
      properties:
        targetAddress:
          type: string
          minLength: 1
          description: On-chain address of the entity being slashed
          example: GDUKMGUGDZQK6YHYA5Z6AY2G4XDSZPSZ3SW5UN3ARVMO6QSRDWP5YLEX
        reason:
          type: string
          minLength: 1
          description: Reason for requesting the slash
          example: Repeated SLA violations on delivery commitments
        requestedBy:
          type: string
          minLength: 1
          description: Identifier of the requester
          example: validator-12
        threshold:
          type: integer
          minimum: 1
          description: "Approve votes required to pass (default: 3)"
          example: 3
        totalSigners:
          type: integer
          minimum: 1
          description: "Total eligible signers (default: 5)"
          example: 5
      required:
        - targetAddress
        - reason
        - requestedBy
    SlashRequestsListResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          type: array
          items:
            $ref: "#/components/schemas/SlashRequest"
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 1
        hasNext:
          type: boolean
          example: false
      required:
        - success
        - data
        - page
        - limit
        - total
        - hasNext
    VoteResult:
      type: object
      properties:
        slashRequestId:
          type: string
          example: a1b2c3d4e5f6a7b8
        voterId:
          type: string
          example: validator-3
        choice:
          $ref: "#/components/schemas/VoteChoice"
        approveCount:
          type: integer
          example: 2
        rejectCount:
          type: integer
          example: 0
        status:
          $ref: "#/components/schemas/SlashRequestStatus"
      required:
        - slashRequestId
        - voterId
        - choice
        - approveCount
        - rejectCount
        - status
    SubmitVoteBody:
      type: object
      properties:
        voterId:
          type: string
          minLength: 1
          example: validator-3
        choice:
          $ref: "#/components/schemas/VoteChoice"
      required:
        - voterId
        - choice
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 5f8d0d55-1c1b-4e9a-9b1a-4e6f6f6f6f6f
        filedBy:
          type: string
          example: GDUKMGUGDZQK6YHYA5Z6AY2G4XDSZPSZ3SW5UN3ARVMO6QSRDWP5YLEX
        respondent:
          type: string
          example: GBVFLWXYZ6JJ7TUSU6QDJ6DOY4J5G5VHGJWSCMHL7QSAHRDEQU3EXFW2
        reason:
          type: string
          example: Goods delivered did not match the agreed specification
        evidence:
          type: array
          items:
            type: string
          example:
            - ipfs://bafybeih...
        status:
          $ref: "#/components/schemas/DisputeStatus"
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T09:00:00.000Z
        deadline:
          type: string
          format: date-time
          example: 2024-01-16T09:00:00.000Z
        resolution:
          type: string
          nullable: true
          example: null
      required:
        - id
        - filedBy
        - respondent
        - reason
        - evidence
        - status
        - createdAt
        - deadline
        - resolution
    DisputeStatus:
      type: string
      enum:
        - pending
        - under_review
        - resolved
        - dismissed
        - expired
    DisputeError:
      type: object
      properties:
        error:
          type: string
          example: NotFound
        message:
          type: string
          example: Dispute not found
      required:
        - error
        - message
    SubmitDisputeBody:
      type: object
      properties:
        filedBy:
          type: string
          pattern: ^G[A-Z2-7]{55}$
          description: Stellar address of the party filing the dispute
          example: GDUKMGUGDZQK6YHYA5Z6AY2G4XDSZPSZ3SW5UN3ARVMO6QSRDWP5YLEX
        respondent:
          type: string
          pattern: ^G[A-Z2-7]{55}$
          description: Stellar address of the respondent
          example: GBVFLWXYZ6JJ7TUSU6QDJ6DOY4J5G5VHGJWSCMHL7QSAHRDEQU3EXFW2
        reason:
          type: string
          minLength: 10
          description: Reason for the dispute (minimum 10 characters)
          example: Goods delivered did not match the agreed specification
        evidence:
          type: array
          items:
            type: string
          minItems: 1
          description: Evidence references (URIs, hashes, etc.) supporting the dispute
          example:
            - ipfs://bafybeih...
        deadlineMs:
          type: integer
          minimum: 3600000
          maximum: 2592000000
          description: Resolution deadline expressed as milliseconds from now (min 1 hour,
            max 30 days)
          example: 86400000
      required:
        - filedBy
        - respondent
        - reason
        - evidence
        - deadlineMs
    DisputeTransitionError:
      type: object
      properties:
        error:
          type: string
          example: Invalid dispute state transition
        code:
          type: string
          example: invalid_dispute_transition
        error_code:
          type: string
          example: invalid_dispute_transition
        message:
          type: string
          example: Invalid transition from "resolved" to "resolved"
      required:
        - error
        - code
        - error_code
        - message
    ResolveDisputeBody:
      type: object
      properties:
        resolution:
          type: string
          minLength: 1
          description: Resolution text describing the outcome
          example: Respondent agreed to a partial refund of 20%
      required:
        - resolution
    DismissDisputeBody:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          description: Reason the dispute is being dismissed
          example: Insufficient evidence provided
      required:
        - reason
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key sent as `Authorization: Bearer <key>`"
...
paths:
...
  /api/governance/slash-requests:
    post:
      summary: Create a slash request
      description: Opens a new slash request awaiting governance votes.
      tags:
        - Governance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlashRequestBody"
      responses:
        "201":
          description: Slash request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlashRequest"
        "400":
          description: Validation error (e.g. threshold < 1, or totalSigners < threshold)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
    get:
      summary: List slash requests
      description: Returns a paginated list of slash requests, optionally filtered by
        status.
      tags:
        - Governance
      security:
        - bearerAuth: []
      parameters:
        - schema:
            $ref: "#/components/schemas/SlashRequestStatus"
          required: false
          name: status
          in: query
        - schema:
            type: integer
            minimum: 1
            example: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 20
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
          required: false
          name: offset
          in: query
      responses:
        "200":
          description: Paginated list of slash requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlashRequestsListResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
  /api/governance/slash-requests/{id}:
    get:
      summary: Get a slash request
      description: Returns a single slash request by ID.
      tags:
        - Governance
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: Slash request ID
            example: a1b2c3d4e5f6a7b8
          required: true
          description: Slash request ID
          name: id
          in: path
      responses:
        "200":
          description: Slash request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlashRequest"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
        "404":
          description: No slash request with this ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
  /api/governance/slash-requests/{id}/votes:
    post:
      summary: Vote on a slash request
      description: Casts an approve/reject vote on a pending slash request.
      tags:
        - Governance
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            description: Slash request ID
            example: a1b2c3d4e5f6a7b8
          required: true
          description: Slash request ID
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVoteBody"
      responses:
        "201":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
        "404":
          description: No slash request with this ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
        "409":
          description: Request already resolved, or voter has already voted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceError"
  /api/disputes:
    post:
      summary: Submit a dispute
      description: Files a new dispute between two Stellar addresses with supporting
        evidence.
      tags:
        - Disputes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDisputeBody"
      responses:
        "201":
          description: Dispute submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "400":
          description: Validation error (invalid Stellar address, reason too short,
            missing evidence, or deadline out of range)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
  /api/disputes/{id}:
    get:
      summary: Get a dispute
      description: Returns a single dispute by ID.
      tags:
        - Disputes
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            description: Dispute ID
            example: 5f8d0d55-1c1b-4e9a-9b1a-4e6f6f6f6f6f
          required: true
          description: Dispute ID
          name: id
          in: path
      responses:
        "200":
          description: Dispute found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
        "404":
          description: No dispute with this ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
  /api/disputes/{id}/review:
    post:
      summary: Mark a dispute under review
      description: Transitions a pending dispute to `under_review`.
      tags:
        - Disputes
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            description: Dispute ID
            example: 5f8d0d55-1c1b-4e9a-9b1a-4e6f6f6f6f6f
          required: true
          description: Dispute ID
          name: id
          in: path
      responses:
        "200":
          description: Dispute marked under review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
        "422":
          description: Invalid state transition (e.g. dispute is not pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeTransitionError"
  /api/disputes/{id}/resolve:
    post:
      summary: Resolve a dispute
      description: Transitions a pending or under-review dispute to `resolved` with a
        resolution note.
      tags:
        - Disputes
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            description: Dispute ID
            example: 5f8d0d55-1c1b-4e9a-9b1a-4e6f6f6f6f6f
          required: true
          description: Dispute ID
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
        "422":
          description: Invalid state transition, missing resolution text, or dispute has
            expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeTransitionError"
  /api/disputes/{id}/dismiss:
    post:
      summary: Dismiss a dispute
      description: Transitions a pending or under-review dispute to `dismissed` with a
        reason.
      tags:
        - Disputes
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            description: Dispute ID
            example: 5f8d0d55-1c1b-4e9a-9b1a-4e6f6f6f6f6f
          required: true
          description: Dispute ID
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DismissDisputeBody"
      responses:
        "200":
          description: Dispute dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeError"
        "422":
          description: Invalid state transition or missing dismiss reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeTransitionError"
...